requirements.txt
# =============================================================================
# OPERADORA DIGITAL DO FUTURO - Dependencias Python
# =============================================================================

# -----------------------------------------------------------------------------
# CAMUNDA EXTERNAL TASK CLIENT
# -----------------------------------------------------------------------------
camunda-external-task-client-python3==4.3.0

# -----------------------------------------------------------------------------
# ORACLE DATABASE (Tasy ERP)
# -----------------------------------------------------------------------------
oracledb==2.0.1

# -----------------------------------------------------------------------------
# HTTP REQUESTS (APIs REST)
# -----------------------------------------------------------------------------
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# -----------------------------------------------------------------------------
# ENVIRONMENT & CONFIGURATION
# -----------------------------------------------------------------------------
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# -----------------------------------------------------------------------------
# DATA PROCESSING
# -----------------------------------------------------------------------------
pandas==2.1.4
numpy==1.26.3

# -----------------------------------------------------------------------------
# MACHINE LEARNING (Estratificacao de Risco)
# -----------------------------------------------------------------------------
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2

# -----------------------------------------------------------------------------
# OPENAI / LLM (GPT-4 para NLP)
# -----------------------------------------------------------------------------
openai==1.6.1
tiktoken==0.5.2

# -----------------------------------------------------------------------------
# REDIS CACHE
# -----------------------------------------------------------------------------
redis==5.0.1

# -----------------------------------------------------------------------------
# ASYNC & CONCURRENCY
# -----------------------------------------------------------------------------
asyncio==3.4.3

# -----------------------------------------------------------------------------
# DATE & TIME
# -----------------------------------------------------------------------------
python-dateutil==2.8.2
pytz==2023.3.post1

# -----------------------------------------------------------------------------
# LOGGING & MONITORING
# -----------------------------------------------------------------------------
structlog==23.3.0

# -----------------------------------------------------------------------------
# TESTING
# -----------------------------------------------------------------------------
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
responses==0.24.1
httpx-mock==0.1.0

# -----------------------------------------------------------------------------
# CODE QUALITY
# -----------------------------------------------------------------------------
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# -----------------------------------------------------------------------------
# UTILITIES
# -----------------------------------------------------------------------------
tenacity==8.2.3
python-json-logger==2.0.7
